fix: race condition when (un-)flattening dirs quickly

+++ src/loader.h
@@ -15,8 +15,8 @@ declare_hmap(previewcache, zsview, Preview *);
 typedef struct Loader {
   dircache dc;
   previewcache pc;
-  usize dir_cache_version; // number of times the cache has been dropped
-  usize preview_cache_version;
+  u64 dir_cache_version; // number of times the cache has been dropped
+  u64 preview_cache_version;
 
   list_loader_timer dir_timers;
   list_loader_timer preview_timers;
